+++ android/src/main/java/com/smileidentity/react/results/BiometricKycCaptureResult.kt
@@ -1,0 +1,9 @@
+package com.smileidentity.react.results
+
+import java.io.File
+
+data class BiometricKycCaptureResult(
+  val selfieFile: File? = null,
+  val livenessFiles: List<File>? = null,
+  val didSubmitBiometricKycJob: Boolean? = null
+)